Cleanup conditional structures
The conditional structures previously used were based on 'case'
statements in a situation where a simpler and more readable `if`
statement would suffice.

This commit clean up the code by switching to a default of `undef` for
the `tags` parameter in the three defined types. This allows a simple
`if` statement to run against the truthyness of the parameter. In
general this should be far more readable.